requirements.txt
transformers==4.28.1
Pillow==9.1.0
torch==1.12.1
